warnings: adopt welder's strict set on every wowlib target
The -Wall/-Wextra lock proved too coarse — the 1191-warning pileup was
only visible through welder's generator TU because ITS flags carry
-Wsign-conversion and ours did not. Now the wowlib targets compile under
welder's full strict set (-Wpedantic -Wshadow -Wconversion
-Wsign-conversion -Wcast-qual -Wold-style-cast -Wnon-virtual-dtor
-Woverloaded-virtual -Wdouble-promotion -Wformat=2
-Wimplicit-fallthrough -Wuseless-cast -Wextra-semi
-Wmisleading-indentation -Wredundant-decls), so anything welder's TUs
would flag, ours flag first — and CI's -Werror makes it fatal.

The sweep's 304 findings, all resolved: 4 real shadows (locals named
"id" over the db::id annotation object, a nested filesystem error_code;
renamed in the .cpp definitions only, so welded keyword names keep
"id"), CascLib's CASC_FILE_DATA_ID macro respelled without its C casts
plus DWORD-vs-uint32_t useless-casts dropped (implicit same-width is
warning-free on both ports), a blp int/uint32 mix, three test-side
index/instantiation pedantries, stb_dxt.h gone SYSTEM (third-party in
our TU), and one deliberate exception documented in-tree: wowlib_py
compiles -Wno-shadow because gcc-16 reports a "template for" loop
variable as shadowing ITSELF once per expansion — 215 bogus warnings
across the binding TUs' 16 enumerator loops, zero real ones.

Full tree: zero warnings under -Werror; 207 ctest + 146 pytest green.

+++ tests/unit/test_listfile_concurrency.cpp
@@ -50,7 +50,7 @@ TEST_CASE("concurrent lookups and registrations keep the maps consistent",
     });
 
   for (int w = 0; w < writers; ++w)
-    threads[w].join();
+    threads[static_cast<std::size_t>(w)].join();
   stop = true;
   threads.clear();
